+++ data/src/main/java/com/weit2nd/data/di/AuthModule.kt
@@ -2,7 +2,7 @@ package com.weit2nd.data.di
 
 import com.weit2nd.data.interceptor.AuthInterceptor
 import com.weit2nd.data.interceptor.LoginInterceptor
-import com.weit2nd.data.source.auth.AuthDataSource
+import com.weit2nd.data.source.token.TokenDataSource
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -14,19 +14,19 @@ import javax.inject.Singleton
 object AuthModule {
     @Singleton
     @Provides
-    fun providesLoginInterceptor(dataSource: AuthDataSource): LoginInterceptor {
+    fun providesLoginInterceptor(dataSource: TokenDataSource): LoginInterceptor {
         return LoginInterceptor(dataSource)
     }
 
     @Singleton
     @Provides
-    fun providesAuthInterceptor(dataSource: AuthDataSource): AuthInterceptor {
+    fun providesAuthInterceptor(dataSource: TokenDataSource): AuthInterceptor {
         return AuthInterceptor(dataSource)
     }
 
     @Singleton
     @Provides
-    fun providesAuthDataSource(): AuthDataSource {
-        return AuthDataSource()
+    fun providesTokenDataSource(): TokenDataSource {
+        return TokenDataSource()
     }
 }
