+++ data/src/main/java/com/weit2nd/data/di/LoginModule.kt
@@ -2,7 +2,7 @@ package com.weit2nd.data.di
 
 import com.weit2nd.data.repository.login.LoginRepositoryImpl
 import com.weit2nd.data.service.LoginService
-import com.weit2nd.data.source.auth.AuthDataSource
+import com.weit2nd.data.source.token.TokenDataSource
 import com.weit2nd.data.source.login.LoginDataSource
 import com.weit2nd.data.util.ActivityProvider
 import com.weit2nd.domain.repository.login.LoginRepository
@@ -20,12 +20,12 @@ object LoginModule {
     @Provides
     fun providesLoginRepository(
         loginDataSource: LoginDataSource,
-        authDataSource: AuthDataSource,
+        tokenDataSource: TokenDataSource,
         activityProvider: ActivityProvider,
     ): LoginRepository {
         return LoginRepositoryImpl(
             loginDataSource = loginDataSource,
-            authDataSource = authDataSource,
+            tokenDataSource = tokenDataSource,
             activityProvider = activityProvider,
         )
     }
